Kernel: Split the TmpFS core files into smaller components
